Migrate the rpmbuild integration test to PyTest
@dmurphy18 can you please revisit this test. It's seems to be failing on
this error:
```
gpg: skipped "SaltStack Packaging Team (This key is for testing package signing only) <packaging@saltstack.com>": No secret key
gpg: signing failed: No secret key
Pass phrase check failed or gpg key expired
Passed invalid arguments: Signing file /tmp/pytest-of-root/pytest-10/test_make_repo_without_gpg_age0/repo/hello-2.10-1.el7.x86_64.rpm failed with proc.status 1.

Usage:
```

FYI, for the test to run on CI we also need https://github.com/saltstack/salt-jenkins/pull/1685 merged and a new batch of golden images
