+++ tests/integration/files/pillar/base/top.sls
@@ -3,7 +3,6 @@ base:
     - generic
     - include
     - glob_include
-    - packagingtest
   'sub_minion':
     - generic
     - sub
